requirements.txt
# Python 3.9+ required
ag2
python-dotenv
yfinance
matplotlib
pandas
numpy
plotly
